refactor: group ferrowl crate-root modules into session/ and cli/
Stage 5 of #89. session_sim + session_e2e_tests -> session/{sim,e2e_tests};
cli + headless -> cli/{mod,headless} (headless is the run-subcommand executor).
Crate root drops from 10 loose files to 6.
